[oc_stormwater/#41] Add map to homepage Rename "ProjectLocations" to "Jurisdictions" where applicable
