+++ Database/ReleaseScript/007 - Improve OrganizationType LegendColor for Local.sql
@@ -1,0 +1,3 @@
+UPDATE dbo.OrganizationType
+SET LegendColor = '#4881cb'
+WHERE OrganizationTypeName = 'Local'
