+++ Source/Neptune.Web/Global.asax.cs
@@ -49,7 +49,7 @@ namespace Neptune.Web
                 "~/Views/Shared/ExpenditureAndBudgetControls/{0}.cshtml",
                 "~/Views/Shared/PerformanceMeasureControls/{0}.cshtml",
                 "~/Views/Shared/ProjectControls/{0}.cshtml",
-                "~/Views/Shared/ProjectLocationControls/{0}.cshtml",
+                "~/Views/Shared/JurisdictionControls/{0}.cshtml",
                 "~/Views/Shared/ProjectWatershedControls/{0}.cshtml",
                 "~/Views/Shared/ProjectUpdateDiffControls/{0}.cshtml"
             } };
